UI update for create_group_screen, UI Started for create_recipe_screen
