+++ rankmyroast/lib/screens/navigational_base_screen/views/groups/screens/widgets/group_member_list_tile.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:rankmyroast/models/group_member.dart';
-import 'package:rankmyroast/screens/navigational_base_screen/views/groups/widgets/screens/widgets/classes/security_level_to_role.dart';
+import 'package:rankmyroast/screens/navigational_base_screen/views/groups/screens/widgets/classes/security_level_to_role.dart';
 import 'package:rankmyroast/services/supabase_helper.dart';
 
 class GroupMemberListTile extends StatefulWidget {
